requirements.txt
streamlit

# AGENT
groq
openai 
agentops
multion
mem0ai